Introduce a log file to track generated bitmap packs
When the `bitmap generation` action is called it results in a new
bitmap (pack and index files) being generated. Collect its name to a
log file (named `.ghs-packs.log` under `ojects/pack` dir) so that
`prune` action (that is going to be introduced) could be used to move
them to the `preserved` dir. And eventually older files (bitmap, pack
and index) could be removed helping to:
* speedup search-for-reuse phase
* conserve the disk space

Note that pack names (in their binary representation of 20 bytes) are
added to the lock safely (with system level file lock) to prevent
concurrent write issues.

Related-to: #157
Change-Id: Ie2eaac0b4a65f2882199c72fa8d1e2475fffdb5f