Introduce a log file to track generated bitmap packs
When the `bitmap generation` action is called it results in a new
bitmap being generated. Collect its name to a log file (named
`packs.log` under `ojects/pack` dir) so that `prune` action (that is
going to be introduced) could be used to move them to the `preserved`
dir. And eventually older bitmap files could be removed (helping to
conserve the disk space).

Note that pack names are added to the lock safely (with system level
file lock) to prevent concurrent write issues.

Bug: Issue 157
Change-Id: Ie2eaac0b4a65f2882199c72fa8d1e2475fffdb5f